+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+### Version 1.32.0
+
+- Splash screen and icon generation now uses @capacitor/assets instead of cordova-res
+- Fix for Capacitor migration recommended replacement plugins
+
 ### Version 1.31.0
 
 - Fix for starter projects switching to project view